[feat] 메인페이지 데이터 출력 기능 itemApi/itemSlice로 이동
