+++ src/features/orderSlice.js
@@ -1,6 +1,6 @@
 // src/features/orderSlice.js
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit'
-import { createOrder, getOrders, getOrderById, cancelOrder, updateOrderStatus, fetchAllOrders, fetchSortData } from '../api/orderApi'
+import { createOrder, getOrders, getOrderById, cancelOrder, updateOrderStatus, fetchAllOrders } from '../api/orderApi'
 
 // 주문 생성 Thunk
 export const createOrderThunk = createAsyncThunk('order/createOrder', async (orderData, { rejectWithValue }) => {
@@ -61,17 +61,6 @@ export const fetchAllOrdersThunk = createAsyncThunk('order/fetchAllOrders', asyn
       return response.data
    } catch (error) {
       return rejectWithValue(error.response?.data?.message || '주문 조회 실패')
-   }
-})
-
-// 회원용 조건부 조회 Thunk
-export const fetchSortDataThunk = createAsyncThunk('order/fetchSortData', async (limit, { rejectWithValue }) => {
-   try {
-      const response = await fetchSortData(limit)
-      console.log('🎈response.data:', response.data)
-      return response.data
-   } catch (error) {
-      return rejectWithValue(error.response?.data?.message || '데이터 조회 실패')
    }
 })
 
@@ -173,19 +162,6 @@ const orderSlice = createSlice({
             state.orders = action.payload.orders
          })
          .addCase(fetchAllOrdersThunk.rejected, (state, action) => {
-            state.loading = false
-            state.error = action.payload
-         })
-         // 회원용 데이터 조회
-         .addCase(fetchSortDataThunk.pending, (state) => {
-            state.loading = true
-            state.error = null
-         })
-         .addCase(fetchSortDataThunk.fulfilled, (state, action) => {
-            state.loading = false
-            state.orders = action.payload
-         })
-         .addCase(fetchSortDataThunk.rejected, (state, action) => {
             state.loading = false
             state.error = action.payload
          })
