+++ src/pages/MainPage.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react'
 import './css/Main.css'
 import { useDispatch, useSelector } from 'react-redux'
-import { fetchSortDataThunk } from '../features/orderSlice'
+import { fetchSortDataThunk } from '../features/itemSlice'
 function MainPage() {
    const dispatch = useDispatch()
-   const { orders: mainData, loading, error } = useSelector((state) => state.order)
+   const { items: mainData, loading, error } = useSelector((state) => state.item)
 
    useEffect(() => {
       dispatch(fetchSortDataThunk(5))
